changed StrictTyper to use double instead of float; added new test to confirm correctness

+++ java/src/net/marcesher/CFStrictTyper.java
@@ -59,9 +59,9 @@ public class CFStrictTyper implements Typer {
 			}
 			
 			try {
-				return Float.parseFloat(stringValue);
+				return Double.parseDouble(stringValue);
 			} catch (Exception e) {
-				//nothing; it's not a float
+				//nothing; it's not a double
 			}
 		}
 		return value;
